refactor(types)!: self-contained generated module; base Column/Filter props generated
The envelope machinery (Node/ColumnNode/FilterNode, ResolveProps, the
augmentable interfaces) cannot be generated from PHP — generics,
conditional resolution, and interface merge seams have no reflectable
counterpart — but it now lives WITH the generator: NodeModuleWriter
inlines it from stubs/envelopes.ts, generated.ts carries zero imports,
and core/types, table/types, and effects/registry re-export from it, so
every dependency arrow points hand-written → generated.

The genuinely shape-like halves ARE generated now: the abstract Column
and Filter bases are transformed directly, replacing the
CommonColumnProps Pick-hack, and the filter fallback gains the
generated label the wire always carries.
