+++ resources/js/core/types.ts
@@ -1,9 +1,17 @@
 import type { ComponentType as ReactComponentType, ReactNode } from "react";
 import type {
-  ComponentPropsMap,
+  CommonNodeProps,
+  ComponentProps,
+  Node,
+  NodeOfType,
+  NodeProps,
   NodeType,
+  NodeUnionOf,
   Option,
   PageContainer as KnownPageContainer,
+  PropsOf,
+  ResolveProps,
+  Schema,
   WireNode,
 } from "@lattice-php/lattice/types/generated";
 import type {
@@ -12,86 +20,21 @@ import type {
   PagePayload as GeneratedPagePayload,
 } from "@lattice-php/lattice/types/generated";
 
-export type { KnownPageContainer, NodeType, Option, WireNode };
-
-export type NodeProps = Record<string, unknown>;
-
-export type CommonNodeProps = {
-  dataBindings?: Record<string, string> | null;
-  hideWhenCollapsed?: boolean | null;
-};
-
-/**
- * Augmentable map of wire `type` string → props for consumer components. Apps
- * extend it via `declare module "@lattice-php/lattice"` so their custom components
- * gain strong prop types; the entries are generated by `php artisan lattice:typescript`.
- * Built-in components resolve through the generated `ComponentPropsMap` below, the
- * same way `ColumnProps` augments `ColumnPropsMap`. Types in neither fall back to a
- * loose bag.
- */
-export interface ComponentProps {}
-
-/**
- * Resolves a `type` string to its payload: consumer augmentations (`TAugment`)
- * win over generated built-ins (`TBuiltins`), then a loose fallback. The one
- * resolver behind the component, column and effect maps.
- */
-export type ResolveProps<
-  TAugment,
-  TBuiltins,
-  TType extends string,
-  TFallback,
-> = TType extends keyof TAugment
-  ? TAugment[TType]
-  : TType extends keyof TBuiltins
-    ? TBuiltins[TType]
-    : TFallback;
-
-export type PropsOf<TType extends string> = ResolveProps<
+export type {
+  CommonNodeProps,
   ComponentProps,
-  ComponentPropsMap,
-  TType,
-  NodeProps
-> &
-  CommonNodeProps;
-
-/**
- * The loose wire node an unknown/untyped `type` resolves to: the shape every node
- * shares, with a loose `props` bag. The single source of the wire-node shape;
- * generated `WireNode` aliases `Node` (i.e. `Node<string>`), which is this.
- */
-type LooseNode = {
-  id?: string;
-  key?: string;
-  type: string;
-  props?: NodeProps;
-  schema?: Schema;
+  KnownPageContainer,
+  Node,
+  NodeOfType,
+  NodeProps,
+  NodeType,
+  NodeUnionOf,
+  Option,
+  PropsOf,
+  ResolveProps,
+  Schema,
+  WireNode,
 };
-
-/**
- * Built-in `props` is required, not optional — the wire serializes the full prop
- * object, so reads need no optional chaining. Unknown types fall back to `LooseNode`.
- */
-export type NodeOfType<TType extends string = string> = string extends TType
-  ? LooseNode
-  : {
-      id?: string;
-      key?: string;
-      props: PropsOf<TType>;
-      schema?: Schema;
-      type: TType;
-    };
-
-export type Node<TType extends string = string> = NodeOfType<TType>;
-
-/**
- * Distributes so `NodeUnionOf<"a" | "b">` is `NodeOfType<"a"> | NodeOfType<"b">`
- * — a discriminated union where `node.type` narrows `node.props`. Domains build
- * their node union from a generated `…NodeType` string union with this.
- */
-export type NodeUnionOf<TTypes extends string> = TTypes extends string ? NodeOfType<TTypes> : never;
-
-export type Schema = Node[];
 
 /** Its `schema` holds exactly one Outlet node. */
 export type LayoutPayload = PageLayoutPayload;
