+++ src/Support/TypeScript/NodeModuleWriter.php
@@ -10,12 +10,11 @@ use Spatie\TypeScriptTransformer\Transformed\Transformed;
 use Spatie\TypeScriptTransformer\Writers\FlatModuleWriter;
 
 /**
- * The base-module writer, plus the imports the generated module can't declare
- * itself: the augmentable envelopes. Component-typed props generate as
- * `Node<"type">`/`Node` (see {@see ComponentTransformer}), Column/Filter-typed
- * props as `ColumnNode`/`FilterNode`, and `WireNode` aliases `Node` — the flat
- * module references those envelopes but never defines them. The resulting
- * import cycles are type-only, so they erase at runtime.
+ * The base-module writer, plus the envelope machinery the generated shapes plug
+ * into (`Node`, `ColumnNode`, `FilterNode`, the `ResolveProps` calculus and the
+ * augmentable interfaces), inlined from the stubs directory so the module is
+ * fully self-contained — zero imports, and hand-written modules re-export from
+ * it rather than the other way around.
  */
 final class NodeModuleWriter extends FlatModuleWriter
 {
@@ -28,33 +27,16 @@ final class NodeModuleWriter extends FlatModuleWriter
     {
         $this->guardUniqueNames($transformed);
 
+        $envelopes = (string) file_get_contents(__DIR__.'/stubs/envelopes.ts');
+
         return array_map(
             fn (WriteableFile $file): WriteableFile => new WriteableFile(
                 $file->path,
-                $this->header($file->contents).$file->contents,
+                $envelopes.$file->contents,
                 $file->changed,
             ),
             parent::output($transformed, $transformedCollection),
         );
-    }
-
-    private function header(string $contents): string
-    {
-        $header = 'import type { Node } from "@lattice-php/lattice/core/types";'."\n";
-
-        $envelopes = array_values(array_filter(
-            ['ColumnNode', 'FilterNode'],
-            fn (string $envelope): bool => preg_match('/\b'.$envelope.'\b/', $contents) === 1,
-        ));
-
-        if ($envelopes !== []) {
-            $header .= sprintf(
-                'import type { %s } from "@lattice-php/lattice/table/types";'."\n",
-                implode(', ', $envelopes),
-            );
-        }
-
-        return $header;
     }
 
     /**
