+++ src/Support/TypeScript/stubs/envelopes.ts
@@ -1,0 +1,157 @@
+// ─── Envelope machinery ─────────────────────────────────────────────────────
+// Inlined verbatim from src/Support/TypeScript/stubs/envelopes.ts by
+// NodeModuleWriter. This is the client-side type calculus the generated shapes
+// below plug into — generics, conditional resolution, and the consumer
+// augmentation seams have no PHP counterpart to generate from, so the
+// generator owns them here and the module stays import-free.
+
+export type NodeProps = Record<string, unknown>;
+
+export type CommonNodeProps = {
+  dataBindings?: Record<string, string> | null;
+  hideWhenCollapsed?: boolean | null;
+};
+
+/**
+ * Augmentable map of wire `type` string → props for consumer components. Apps
+ * extend it via `declare module "@lattice-php/lattice"` so their custom components
+ * gain strong prop types; the entries are generated by `php artisan lattice:typescript`.
+ * Built-in components resolve through the generated `ComponentPropsMap` below.
+ * Types in neither fall back to a loose bag.
+ */
+export interface ComponentProps {}
+
+/** The column counterpart of `ComponentProps`. */
+export interface ColumnProps {}
+
+/** The filter counterpart of `ComponentProps`. */
+export interface FilterProps {}
+
+/** The effect counterpart of `ComponentProps`. */
+export interface EffectProps {}
+
+/**
+ * Resolves a `type` string to its payload: consumer augmentations (`TAugment`)
+ * win over generated built-ins (`TBuiltins`), then a loose fallback. The one
+ * resolver behind the component, column, filter and effect maps.
+ */
+export type ResolveProps<
+  TAugment,
+  TBuiltins,
+  TType extends string,
+  TFallback,
+> = TType extends keyof TAugment
+  ? TAugment[TType]
+  : TType extends keyof TBuiltins
+    ? TBuiltins[TType]
+    : TFallback;
+
+export type PropsOf<TType extends string> = ResolveProps<
+  ComponentProps,
+  ComponentPropsMap,
+  TType,
+  NodeProps
+> &
+  CommonNodeProps;
+
+/**
+ * The loose wire node an unknown/untyped `type` resolves to: the shape every node
+ * shares, with a loose `props` bag. The single source of the wire-node shape;
+ * generated `WireNode` aliases `Node` (i.e. `Node<string>`), which is this.
+ */
+type LooseNode = {
+  id?: string;
+  key?: string;
+  type: string;
+  props?: NodeProps;
+  schema?: Schema;
+};
+
+/**
+ * Built-in `props` is required, not optional — the wire serializes the full prop
+ * object, so reads need no optional chaining. Unknown types fall back to `LooseNode`.
+ */
+export type NodeOfType<TType extends string = string> = string extends TType
+  ? LooseNode
+  : {
+      id?: string;
+      key?: string;
+      props: PropsOf<TType>;
+      schema?: Schema;
+      type: TType;
+    };
+
+export type Node<TType extends string = string> = NodeOfType<TType>;
+
+/**
+ * Distributes so `NodeUnionOf<"a" | "b">` is `NodeOfType<"a"> | NodeOfType<"b">`
+ * — a discriminated union where `node.type` narrows `node.props`. Domains build
+ * their node union from a generated `…NodeType` string union with this.
+ */
+export type NodeUnionOf<TTypes extends string> = TTypes extends string ? NodeOfType<TTypes> : never;
+
+export type Schema = Node[];
+
+/**
+ * The generated `Column` base props (serialized by every column regardless of
+ * type) back the fallback, so even an unaugmented custom column keeps its
+ * common concerns typed.
+ */
+export type ColumnPropsOf<TType extends string> = ResolveProps<
+  ColumnProps,
+  ColumnPropsMap,
+  TType,
+  Column & Record<string, unknown>
+>;
+
+export type CommonColumnProps = Column;
+
+/**
+ * A column node, authored and consumed like a field/component: `key`/`type`
+ * stay top-level, every common concern (`label`/`width`/`filter`/…) lives in
+ * `props` on each generated column type. `schema` is only present for a
+ * container column: the bound component nodes rendered per row. The column
+ * counterpart of `NodeOfType`.
+ */
+export type ColumnNode<TType extends string = string> = {
+  type: TType;
+  key: string;
+  props: ColumnPropsOf<TType>;
+  schema?: Schema;
+};
+
+export type FilterPropsOf<TType extends string> = ResolveProps<
+  FilterProps,
+  FilterPropsMap,
+  TType,
+  Filter & Record<string, unknown>
+>;
+
+export type FilterNodeType = keyof FilterPropsMap & string;
+
+/**
+ * A dedicated table filter, authored and consumed like a field/component:
+ * `key`/`type` stay top-level, `props` carries `label` plus whatever the
+ * filter type contributes. `schema` is the bound field(s) rendering the
+ * filter's control and is absent when empty. The filter counterpart of
+ * `ColumnNode`.
+ */
+export type FilterNode<TType extends string = FilterNodeType> = {
+  type: TType;
+  key: string;
+  props: FilterPropsOf<TType>;
+  schema?: Node[];
+};
+
+export type EffectPropsOf<TType extends string> = ResolveProps<
+  EffectProps,
+  EffectPropsMap,
+  TType,
+  Record<string, unknown>
+>;
+
+export type EffectOf<TType extends string> = string extends TType
+  ? Effect
+  : { type: TType; props: EffectPropsOf<TType> };
+
+// ─── Generated shapes ───────────────────────────────────────────────────────
